Handle errors in display functions more gracefully
Before, when one repr_xxx was raising an error, the complete display machinery
was impacted and no output was displayed. Now only the specific output is
omitted and a user friendly error message is shown on stderr and the stacktrace
(up to now pretty useless because it just shows one line) is shown on the
console.