Send stacktrace on display error
Before, we only showed the error message, but no stacktrace in the notebook. The
stacktrace was shown in the console. Now it is shown in both.

This is in line what the IPython kernel does. We also now only display the
stacktrace for the method which does something wrong and not our internal
execution stuff...