Save plot size in each recordedplot and use it to build images
This fixes the problem that the size of a plot would be taken from the options
when a new plot was created or when the cell was finished and not when the
original plot was created. This resulted in problems when you set the plot size
between two plots in the same cell or more specificly, you would have to set the
plot size *after* you created a plot (and befor the cell was finished or a new
plot was created.

No we save the plot size at the time the plot is created and use that size to
build the images.

```
set_plot_size <- function(width = 7, height = 7){
    options(repr.plot.width = width)
    options(repr.plot.height = height)
}
set_plot_size(3,1) # ignored before, now taken for the first plot
qplot(1:10,1:10)
set_plot_size(3,3) # before for the first plot, now for the second
qplot(1:10,1:10)
set_plot_size(1,3) # before taken for the second plot, now only for following plots
```