Add environment for functions which shadow R functions
We need to get some functions into the search path so that they shadow functions
of R proper. One examples is `quit()` which we have to implement so that it
properly shuts down the kernel and tells the frontend that it wants to shutdown.

After some experimentation, the environment had to be added in the main codepath
so that a) other places in the code could access it and b) it was actually
places into the searchpath between the already loaded packages and the
.Globalenv.