Fix tests
Turns out, we are using jupyter_kernel_tests in ways it was not designed to be
used. So fix that.

I also reorganized the tests to reset options instead of restarting the kernel
each time. This should lead to faster tests. This also meant that I could
reorder the tests so that the basic order is

    "test with normal options -> test with special options"