Bind base quit method to new name to preserve use during shutdown
The replacement of the base `quit()` function with `execution.r/quit()`
prevents message-based shutdown requests from completing, triggering the
juptyer framework's kill sequence.  If the shadowing is not performed,
kernels can be properly shutdown via the initial request and prevent
the additional overhead and potential unreliability of kill sequence
(signals, REST apis, etc.).

Since my knowledge of R is extremely limited, there may be far better
ways to handle this, so please feel free to comment to that affect.
Also, it's not clear to me what functionality I'm preventing by making
this change.

The previous attempt at addressing this was to use a new option to indicate
that shadowing of `quit()` not take place.  However, after discussion
it was felt that somehow preserving the original quit call would be best.
Since the shadowing replaces quit with a custom method, this change
essentially binds a new name `base_quit` to the base method of a new
environment (`backup_env`) prior to it being shadowed (and similar to how
other "original" methods are preserved.  Then, during shutdown requests,
`backup_env$base_quit()` is invoked instead, permitting the kernel to
properly shutdown.

Please refer to Issue #569 for how to test.  The key thing is that,
at a minimum, Notebook DEBUG logging be enabled so one can determine
if the kill sequence is being invoked.  Another way is if the kernel's
termination takes on the order of 5 seconds vs sub-second.

Fixes #569